Update 0x00e2b480 to Color, not text
Based on the usage of this variable within function 0x4503e0, char
values r, g, b, and a are added to this variable.

Function 0x4503e0 is swrText_CreateEntry, and this variable is used to
represent the text color. It was incorrectly labeled as position -
however position is at 0x00e34860.

This was causing a duplicate definition within globals.h