Autogenerate bidirectional hooks on compile
This introduces a script that automatically generates all hooks by
reading through the source code, and looking for methods with a comment
like:

// 0x<address> HOOK
<function declaration>

It must be in this exact format, with the function on a new line. This
will then generate an entry into hook_generated.c that hooks the address
mentioned to that function

For other functions that just have:

// 0x<address> (anything else)
<function declaration>

A reverse hook will be created, hooking from the function itself back
into the executable, so those functions can be used before they are
implemented, allowing for partial decompilation (just a function and not
the functions it calls)

This is an attempt to bring together the work done for the basic hooking
in the launcher, and the dynamic hooking within dinput_hook.

Advantages:
No external depedendencies (no need for detours or BFD)
Generates a file that allows for auditing, so its clear what is getting
hooked
Works in both debug and release builds
Runs automatically every time the source files change
Works for the launcher as well as dinput method for dependency
Maintains a log for hooks

Disadvantages:
Requires running the cmake build after each change to hook, BFD allowed
this to happen magically/automatically without running the build
Does not allow calling the original function when hooked, although
likely this can be implemented